+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changes
 
 ## Unreleased - 2020-xx-xx
-
+### Changed
+* `middleware::normalize::TrailingSlash` enum is now accessible. [#1673]
 
 ## 3.0.0 - 2020-09-11
 * No significant changes from `3.0.0-beta.4`.